// SIC/SIC Admin/src/types/seasonalTopics.ts
export type SeasonalTopicDataArr = {
	id: number;
	uuid: string;
	name: string;
	namePinging: string;
	nameSimplifiedChinese: string;
	nameTraditionalChinese: string;
	image: string;
	rewardStars: number;
	startDate: string;
	endDate: string;
	order: number;
	isActive: boolean;
	createdAt: string;
	updatedAt: string;
	createdBy: number;
};
// ...
export type ActivityKaraokeSeasonalData = {
	onSubmit: () => void;
	onClose: () => void;
	editData?: SeasonalTopicDataArr | null;
};